Separate fixtures into different files.
